+++ c_compiler/driver/float_EQUAL_driver.c
@@ -1,10 +1,10 @@
-float EQUAL(float a, float b, int x);
+int EQUAL(float a, float b, float x);
 
 int main(){
 	float a = 40.456;
 	float b = 2.345;
-	unsigned x = 3;
-	float result = a==b-x;
+	float x = 3;
+	int result = a==b-x;
 	if(result == EQUAL(a,b,x)){
 		return 0;
 	}
