+++ c_compiler/driver/float_NOTEQUAL_driver.c
@@ -1,9 +1,9 @@
-float NOTEQUAL(float a, float b, int x);
+int NOTEQUAL(float a, float b, float x);
 
 int main(){
 	float a = 40.456;
 	float b = 2.345;
-	unsigned x = 3;
+	float x = 3;
 	float result = a!=b-x;
 	if(result == NOTEQUAL(a,b,x)){
 		return 0;
