+++ c_compiler/driver/float_mixed_truncate2.c
@@ -1,5 +1,5 @@
 
-float operations(float a, float b, int x){
-	float s = a!=x==b;
+int operations(float a, float b, float x){
+	int s = a!=x==b;
 	return s;
 }
