+++ c_compiler/driver/float_mixed_truncate4_driver.c
@@ -1,9 +1,9 @@
-float operations(float a, float b, int x);
+float operations(float a, float b, float x);
 
 int main(){
 	float a = 40.456;
 	float b = 2.345;
-	int x = 3;
+	float x = 3;
 	float result = a+x+b;
 	if(result == operations(a,b,x)){
 		return 0;
