+++ c_compiler/driver/float_operations16_driver.c
@@ -3,7 +3,7 @@ float f();
 int main(){
    float x = 3.456;
 	float y = 0.455;
-	unsigned z = 4;
+	float z = 4;
 
 	y=z/y;
