+++ c_compiler/driver/float_operations23_driver.c
@@ -6,7 +6,7 @@ int main(){
   
    float x = 15.88;
 	float k = 1.455;
-	unsigned z = 4;
+	float z = 4;
 	for(z<18;z++;){
 		if(z>=x){
 			break;
