+++ c_compiler/driver/float_trick.c
@@ -1,5 +1,5 @@
 
-float NOTEQUAL(float a, float b, int x){
+float NOTEQUAL(float a, float b, float x){
 
 	return a*b*x;
 }
